fix : TextEstimator enum moved to params.py
